backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Cors==4.0.0
PyMySQL==1.1.0
cryptography==42.0.8
python-dotenv==1.0.1
Pillow>=10.4.0
requests==2.32.3
Werkzeug==3.0.3
